Refactor codegen utilities and node traits structure
Moved codegen utilities for TokenStream conversion from node_codegen.rs to a new codegen.rs file. Renamed node_codegen.rs to node_traits.rs to better reflect its contents, and updated all imports and re-exports accordingly. This improves code organization and separation of concerns between code generation utilities and node trait definitions.
