+++ crates/burn-import/src/burn/mod.rs
@@ -4,18 +4,19 @@ pub mod graph;
 #[macro_use]
 mod registry_macro;
 
-mod node_codegen;
+mod codegen;
 mod node_registry;
 #[cfg(test)]
 pub(crate) mod node_test;
+mod node_traits;
 
 mod imports;
 
 mod scope;
 mod ty;
 
+pub(crate) use codegen::ToTokens;
 pub(crate) use imports::*;
-pub(crate) use node_codegen::ToTokens;
 pub(crate) use scope::*;
 pub(crate) use ty::*;
